backend/pyproject.toml
[project]
name = "agentic-bay-backend"
version = "0.1.0"
description = "Agentic Bay Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "redis[hiredis]>=5.0.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0,<5.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "websockets>=12.0",
    "anyio>=4.4.0",
    "greenlet>=3.0.0",
    "aiohttp>=3.13.5",
    "web3>=7.15.0",
    "loguru>=0.7.3",
    "anthropic>=0.96.0",
    "voyageai>=0.3.7",
    "jwt>=1.4.0",
    "aiosqlite>=0.22.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
]
